+++ spec/injectable.spec.ts
@@ -1,4 +1,4 @@
-import { Injectable, TypeArtisan } from '../src';
+import { Injectable, TypeManager } from '../src';
 
 @Injectable()
 class UserService
@@ -16,8 +16,8 @@ describe('Injectable decorator', () =>
 {
     it('should register a type as injectable', () =>
     {
-        const userServiceMetadata  = TypeArtisan.extractTypeMetadata(UserService);
-        const emailServiceMetadata = TypeArtisan.extractTypeMetadata(EmailService);
+        const userServiceMetadata  = TypeManager.extractTypeMetadata(UserService);
+        const emailServiceMetadata = TypeManager.extractTypeMetadata(EmailService);
 
         expect(userServiceMetadata.typeOptions.injectable).toBeTrue();
         expect(emailServiceMetadata.typeOptions.injectable).toBeFalse();