+++ spec/serializable.spec.ts
@@ -1,4 +1,4 @@
-import { Property, Serializable, Type, TypeArtisan } from '../src';
+import { Property, Serializable, Type, TypeManager } from '../src';
 
 @Type()
 class User
@@ -11,7 +11,7 @@ describe('Serializable decorator', () =>
 {
     it('should register property as serializable', () =>
     {
-        const userMetadata      = TypeArtisan.extractTypeMetadata(User);
+        const userMetadata      = TypeManager.extractTypeMetadata(User);
         const userNameMetadata  = userMetadata.propertyMetadataMap.get('name');
         const userEmailMetadata = userMetadata.propertyMetadataMap.get('email');