+++ spec/use-implicit-conversion.spec.ts
@@ -1,4 +1,4 @@
-import { Property, Type, TypeArtisan, UseImplicitConversion } from '../src';
+import { Property, Type, TypeManager, UseImplicitConversion } from '../src';
 
 @Type()
 @UseImplicitConversion()
@@ -11,14 +11,14 @@ describe('Use implicit conversion decorator', () =>
 {
     it('should enable use of implicit conversion for a type', () =>
     {
-        const userMetadata = TypeArtisan.extractTypeMetadata(User);
+        const userMetadata = TypeManager.extractTypeMetadata(User);
 
         expect(userMetadata.typeOptions.useImplicitConversion).toBeTrue();
     });
 
     it('should enable use of implicit conversion for a property', () =>
     {
-        const userMetadata     = TypeArtisan.extractTypeMetadata(User);
+        const userMetadata     = TypeManager.extractTypeMetadata(User);
         const userNameMetadata = userMetadata.propertyMetadataMap.get('name');
 
         expect(userNameMetadata).toBeDefined();